# api/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: MC Taskmaster
  version: "1.0"
servers:
- url: http://api.task.jazure.net
paths:
  /categories:
    get:
      operationId: get_categories
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Category'
                type: array
          description: OK
      x-openapi-router-controller: openapi_server.controllers.default_controller
    post:
      operationId: create_category
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_category_request'
        description: Create a new category
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Category created successfully
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /categories/{category_id}:
    delete:
      operationId: delete_category
      parameters:
      - explode: false
        in: path
        name: category_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Category deleted successfully
        "404":
          description: Category not found
      x-openapi-router-controller: openapi_server.controllers.default_controller
    get:
      operationId: get_category
      parameters:
      - explode: false
        in: path
        name: category_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Category corresponding to id
        "404":
          description: Category not found
      x-openapi-router-controller: openapi_server.controllers.default_controller
    put:
      operationId: update_category
      parameters:
      - explode: false
        in: path
        name: category_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_category_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Category updated successfully
        "404":
          description: Category not found
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /tasks:
    get:
      operationId: get_tasks
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/get_tasks_200_response'
          description: OK
      x-openapi-router-controller: openapi_server.controllers.default_controller
    post:
      operationId: create_task
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_task_request'
        description: Create a new task
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Task created successfully
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /tasks/{task_id}:
    delete:
      operationId: delete_task
      parameters:
      - explode: false
        in: path
        name: task_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Task deleted successfully
        "404":
          description: Task not found
      x-openapi-router-controller: openapi_server.controllers.default_controller
    get:
      operationId: get_task
      parameters:
      - explode: false
        in: path
        name: task_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Task corresponding to id
        "404":
          description: Task not found
      x-openapi-router-controller: openapi_server.controllers.default_controller
    put:
      operationId: update_task
      parameters:
      - explode: false
        in: path
        name: task_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/create_task_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Task updated successfully
        "404":
          description: Task not found
      x-openapi-router-controller: openapi_server.controllers.default_controller
components:
  schemas:
    Category:
      example:
        name: name
        id: id
      properties:
        id:
          title: id
          type: string
        name:
          title: name
          type: string
      title: Category
    Task:
      example:
        max_session: 1
        description: description
        id: id
        categories:
        - name: name
          id: id
        - name: name
          id: id
        min_session: 6
        enabled: true
        points: 0
      properties:
        id:
          title: id
          type: string
        description:
          title: description
          type: string
        points:
          title: points
          type: integer
        min_session:
          title: min_session
          type: integer
        max_session:
          title: max_session
          type: integer
        enabled:
          title: enabled
          type: boolean
        categories:
          items:
            $ref: '#/components/schemas/Category'
          title: categories
          type: array
      required:
      - categories
      - description
      - id
      - name
      - points
      title: Task
      type: object
    get_tasks_200_response:
      example:
        count: 5
        tasks:
        - max_session: 1
          description: description
          id: id
          categories:
          - name: name
            id: id
          - name: name
            id: id
          min_session: 6
          enabled: true
          points: 0
        - max_session: 1
          description: description
          id: id
          categories:
          - name: name
            id: id
          - name: name
            id: id
          min_session: 6
          enabled: true
          points: 0
      properties:
        tasks:
          items:
            $ref: '#/components/schemas/Task'
          title: tasks
          type: array
        count:
          title: count
          type: integer
      required:
      - count
      - tasks
      title: get_tasks_200_response
      type: object
    create_task_request:
      properties:
        description:
          title: description
          type: string
        points:
          title: points
          type: integer
        min_session:
          title: min_session
          type: integer
        max_session:
          title: max_session
          type: integer
      required:
      - description
      - points
      title: create_task_request
      type: object
    create_category_request:
      properties:
        name:
          title: name
          type: string
      required:
      - name
      title: create_category_request
      type: object